# docs/swagger.yaml
basePath: /
definitions:
  models.CreateRatingPlace:
    properties:
      place_id:
        type: integer
      rating:
        type: integer
      review:
        type: string
      user_id:
        type: number
    type: object
  models.RatingPlace:
    properties:
      created_at:
        type: string
      id:
        type: integer
      place_id:
        type: integer
      rating:
        type: integer
      review:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8001
info:
  contact: {}
  title: Swagger rating services
  version: "1.0"
paths:
  /ratings:
    get:
      consumes:
      - application/json
      parameters:
      - description: getting ratings for place
        in: query
        name: place
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RatingPlace'
            type: array
      summary: Get list of ratings
      tags:
      - ratings
    post:
      consumes:
      - application/json
      parameters:
      - description: Create rating
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateRatingPlace'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RatingPlace'
      summary: Create rating
      tags:
      - ratings
  /ratings/{id}:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/models.RatingPlace'
      summary: Update rating
      tags:
      - ratings
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RatingPlace'
      summary: Get rating by id
      tags:
      - ratings
    patch:
      consumes:
      - application/json
      parameters:
      - description: Update rating
        in: body
        name: input
        schema:
          $ref: '#/definitions/models.CreateRatingPlace'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RatingPlace'
      summary: Update rating
      tags:
      - ratings
securityDefinitions:
  ApiKeyAAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"